requirements.txt
numpy==1.24.3
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.2
accelerate==0.25.0
moviepy==1.0.3
ffmpeg-python==0.2.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
soundfile==0.12.1
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
alembic==1.12.1
pydub==0.25.1
tqdm==4.66.1
sentencepiece==0.1.99
protobuf==4.25.1
psutil==5.9.6
requests==2.31.0
asyncpg==0.29.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0
pytest-env==1.1.1
pytest-mock==3.12.0
pytest-sugar==0.9.7
pytest-xdist==3.3.1
pytest-timeout==2.2.0
pytest-rerunfailures==12.0
pytest-randomly==3.15.0
pytest-picked==0.5.0
pytest-instafail==0.5.0
pytest-html==4.1.1
pytest-benchmark==4.0.0
pytest-aiohttp==1.0.5
pytest-assume==2.4.3 